♻️ Refactor: 컴포넌트 구조 및 함수형 리팩토링
